+++ src/screens/transaction-sum-address.tsx
@@ -1,4 +1,4 @@
-import React, {useCallback, useRef} from 'react';
+import React, {useCallback, useMemo, useRef, useState} from 'react';
 
 import {TransactionAddress} from '@app/components/transaction-address';
 import {app} from '@app/contexts';
@@ -11,9 +11,26 @@ export const TransactionSumAddressScreen = () => {
   const wallets = useWalletsVisible();
   const contacts = useRef(Contact.getAll().snapshot()).current;
 
+  const [address, setAddress] = useState(route.params?.to || '');
+  const filteredWallets = useMemo(() => {
+    if (!wallets || !wallets.length) {
+      return;
+    }
+
+    if (!address) {
+      return wallets.snapshot();
+    }
+
+    return wallets
+      .filtered(
+        `address CONTAINS[c] '${address}' or name CONTAINS[c] '${address}'`,
+      )
+      .snapshot();
+  }, [address, wallets]);
+
   const onDone = useCallback(
-    (address: string) => {
-      app.emit(route.params.event, address);
+    (result: string) => {
+      app.emit(route.params.event, result);
       navigation.goBack();
     },
     [navigation, route.params.event],
@@ -21,9 +38,10 @@ export const TransactionSumAddressScreen = () => {
 
   return (
     <TransactionAddress
-      wallets={wallets}
+      address={address}
+      setAddress={setAddress}
+      filteredWallets={filteredWallets}
       contacts={contacts}
-      initial={route.params.to}
       onAddress={onDone}
     />
   );
